Initial Filtering_(Sine Creator)
The code includes a chebyshev 2 bandpass or notch filtering function including the symmetric padding required for ECoG filtering and uses zero-phase filtering. Outputs plots of filter parameters and results of filtering if desired. Can be used either as notch or chebyshev2 bandpass but not both.
Also a Sine wave creator is added for checking the filter performance.