Finger State/Time Determining
The fingerMovingIndex function takes in a single finger movement signal and returns 4 second windows of times capturing the locations of finger movement roughly. The function plots all the detected finger moving regions on one figure as subplots. 
The Function can be edited or changed to give different window sizes or more specific location detection, more output parameters can be incorporated.
fingerRestingIndex will aim to find all the locations of the resting regions of the finger using the output of the fingerMovingIndex(), however might not be necessary.